api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
bcrypt>=4.1.2
supabase>=2.4.0
python-dotenv>=1.0.1
